requirements.txt
PyYAML==6.0
scapy==2.5.0